AddScheduleDialog コンポーネントから FormControl コンポーネントを抜き出し
