+++ src/components/Navigation/AddScheduleDialog.tsx
@@ -6,12 +6,10 @@ import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import TextField from '@material-ui/core/TextField';
-import InputLabel from '@material-ui/core/InputLabel';
-import FormControl from '@material-ui/core/FormControl';
-import NativeSelect from '@material-ui/core/NativeSelect';
 import Button from '@material-ui/core/Button';
 import ScheduleDatePicker from 'components/Navigation/AddScheduleDialogParts/ScheduleDatePicker';
 import ScheduleTimePicker from 'components/Navigation/AddScheduleDialogParts/ScheduleTimePicker';
+import ScheduleTagSelect from 'components/Navigation/AddScheduleDialogParts/ScheduleTagSelect';
 import type { UseScheduleType } from 'hooks/useSchedules';
 import type { TagBox } from 'hooks/useTags';
 import { useAddSchedule } from 'hooks/useAddSchedule';
@@ -51,22 +49,6 @@ const AddScheduleDialog: VFC<Props> = React.memo<Props>(
         inputHandlers.setScheduleTitle(event.target.value);
       },
       [inputHandlers]
-    );
-
-    const handleTagSelectBox = React.useCallback(
-      (event: ChangeEvent<HTMLSelectElement>) => {
-        const selectedTagBox = tagBoxes.find(
-          (tagBox) => tagBox.tag.name === event.target.value
-        );
-
-        if (selectedTagBox === undefined) {
-          throw new Error();
-        }
-
-        inputHandlers.setSelectedTagName(selectedTagBox.tag.name);
-        inputHandlers.setSelectedTagId(selectedTagBox.tag.tagId);
-      },
-      [tagBoxes, inputHandlers]
     );
 
     const handleSubmit = React.useCallback(() => {
@@ -116,25 +98,12 @@ const AddScheduleDialog: VFC<Props> = React.memo<Props>(
             selectedStartTime={inputValues.selectedStartTime}
             setSelectedStartTime={inputHandlers.setSelectedStartTime}
           />
-          <FormControl>
-            <InputLabel shrink htmlFor="tag">
-              タグ
-            </InputLabel>
-            <NativeSelect
-              inputProps={{ id: 'tag' }}
-              value={inputValues.selectedTagName}
-              onChange={handleTagSelectBox}
-            >
-              <option value="">必ず選択してください</option>
-              {tagBoxes.map((tagBox) => (
-                <>
-                  <option key={tagBox.tag.name} value={tagBox.tag.name}>
-                    {tagBox.tag.name}
-                  </option>
-                </>
-              ))}
-            </NativeSelect>
-          </FormControl>
+          <ScheduleTagSelect
+            selectedTagName={inputValues.selectedTagName}
+            setSelectedTagId={inputHandlers.setSelectedTagId}
+            setSelectedTagName={inputHandlers.setSelectedTagName}
+            tagBoxes={tagBoxes}
+          />
         </DialogContent>
         <DialogActions>
           <Button variant="contained" onClick={closeDialog}>
